+++ QualityCheck/changelog.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## Version 2022042501
+
+* fixing issue when customers don't have sg3_utils installed, moving to lsscsi instead of sg_map
+* adding exception handling for disks that are not used, just attached
+* fixing an issue where Db2 OS/DB combination wasn't working
+
+## Version 2022041301
+
+* adding a check if sg_map is installed
+
+## Version 2022041102
+
+* adding case sensitivity to parameters
+
+## Version 2022041101
+
+* adding sudo check
+* adding version check
+
 ## Version 2022022301
 
 * changing naming to "Quality Check for SAP workloads on Azure"